Change index (date) on time_entry to btree instead of brim and add (employee, date) composite index.
The brim index didn't work properly in production, and caused every row
to be checked even when only looking up a single date. This can happen
with a brim index if rows aren't inserted in order.

The (employee, date) index improves the performance of the
accumulated_hours_for_employee function. It previously had to scan the entire table.